+++ frontend/views/default.jsx
@@ -9,7 +9,7 @@ function Def (html) {
                 <title>RoseDiem Art Duels</title>
                 {/* Bootstrap CDN link */}
                 <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossOrigin="anonymous"/>
-                <link rel='stylesheet' href='/css/style.css'/>
+                <link rel='stylesheet' href='style.css'/>
             </head>
             <body>
                 {/* NAVBAR */}
@@ -17,7 +17,7 @@ function Def (html) {
                     {/* NAVBAR LINKS */}
                     {/* HOME */}
                     <li>
-                        <a href='/'>Home</a>
+                        <a id="navHome" href='/'>RoseDiem Art Duels</a>
                     </li>
                     {/* RULES */}
                     <li>